Implement pretty-printing
The '-v' option now prints rudimentary equation trees along with the
solutions.

Refactor `os` so that it is simpler to obtain a readable string
representation of each `Op`, in order to have something useful to print
in the tree.

Implement optparse-applicative to handle the '-'/"--verbose" option.
Refactor operator list printing to use optparse-applicative while we're
at it.

Use "a:b" instead of "a : b" to uh save horizontal space in the new
code. Sorry.

Signed-off-by: Alexander Berntsen <alexander@plaimi.net>